refactor(lotto): 클래스 위치 변경 및 Lottos 생성자 수정
