+++ src/main/java/lotto/LottoMain.java
@@ -1,0 +1,5 @@
+package lotto;
+
+public class LottoMain {
+
+}
