+++ src/test/java/lotto/RandomLottoNumberGeneratorTest.java
@@ -2,6 +2,8 @@ package lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotto.domain.LottoNumberGenerator;
+import lotto.domain.RandomLottoNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
